use new headers in article layouts
