+++ src/Components/Publishing/Layouts/StandardLayout.tsx
@@ -82,7 +82,7 @@ export class StandardLayout extends React.Component<
                 isTruncated={isTruncated}
                 hideButton={this.removeTruncation}
               >
-                <Header article={article} isMobile={isMobile} />
+                <Header article={article} />
 
                 <StandardLayoutParent>
                   <StandardLayoutContainer>
