+++ src/Components/Publishing/__stories__/ArticleFeature.story.tsx
@@ -93,7 +93,6 @@ story
           display={Display("image")}
           relatedArticlesForPanel={RelatedPanel}
           relatedArticlesForCanvas={RelatedCanvas}
-          emailSignupUrl="#"
           isTruncated
         />
       </ContextProvider>
